[REFACTOR] Component 이벤트 객체 패키지 변경 (#95)
