+++ src/main/java/ject/componote/domain/component/application/ComponentService.java
@@ -7,7 +7,7 @@ import ject.componote.domain.component.dao.ComponentRepository;
 import ject.componote.domain.component.dao.ComponentSummaryDao;
 import ject.componote.domain.component.domain.Component;
 import ject.componote.domain.component.domain.ComponentType;
-import ject.componote.domain.component.dto.find.event.ComponentViewCountIncreaseEvent;
+import ject.componote.domain.component.dto.event.ComponentViewCountIncreaseEvent;
 import ject.componote.domain.component.dto.find.request.ComponentSearchRequest;
 import ject.componote.domain.component.dto.find.response.ComponentDetailResponse;
 import ject.componote.domain.component.dto.find.response.ComponentSummaryResponse;
