+++ src/test/java/ject/componote/domain/component/application/ComponentViewCountEventHandlerTest.java
@@ -3,7 +3,7 @@ package ject.componote.domain.component.application;
 import ject.componote.domain.common.model.Count;
 import ject.componote.domain.component.dao.ComponentRepository;
 import ject.componote.domain.component.domain.Component;
-import ject.componote.domain.component.dto.find.event.ComponentViewCountIncreaseEvent;
+import ject.componote.domain.component.dto.event.ComponentViewCountIncreaseEvent;
 import ject.componote.domain.component.error.NotFoundComponentException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
